# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Polideportivo API",
        "description": "Polideportivo API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/socios": {
            "get": {
                "tags": [
                    "Socios"
                ],
                "summary": "Get all the socios",
                "description": "Returns all socios as a JSON response",
                "operationId": "6df39126d393a0a4421993707731932c",
                "responses": {
                    "200": {
                        "description": "Request successful"
                    }
                }
            },
            "post": {
                "tags": [
                    "Socios"
                ],
                "summary": "Create a new socio",
                "description": "Create a new socio and return it as a JSON response",
                "operationId": "00fb9ef6a0f178c53d02a37da07d41e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Fátima"
                                    },
                                    "edad": {
                                        "type": "integer",
                                        "example": "19"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Socio created successfully"
                    }
                }
            }
        },
        "/api/v1/socios/{socio}": {
            "get": {
                "tags": [
                    "Socios"
                ],
                "summary": "Get a specific socio",
                "description": "Returns a specific socio as a JSON response",
                "operationId": "5009b2d16fe31f6100a3d786dc30dd39",
                "parameters": [
                    {
                        "name": "socio",
                        "in": "path",
                        "description": "ID of the socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful"
                    }
                }
            },
            "put": {
                "tags": [
                    "Socios"
                ],
                "summary": "Update a specific socio",
                "description": "Update a specific socio and return it as a JSON response",
                "operationId": "58e674a5c137dd0a119eed1a237b57e4",
                "parameters": [
                    {
                        "name": "socio",
                        "in": "path",
                        "description": "ID of the socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Fátima"
                                    },
                                    "edad": {
                                        "type": "integer",
                                        "example": "19"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Socios"
                ],
                "summary": "Delete a specific socio",
                "description": "Delete a specific socio and return a JSON response",
                "operationId": "9c3dda2c8f5ca6f339119b87849f6b1c",
                "parameters": [
                    {
                        "name": "socio",
                        "in": "path",
                        "description": "ID of the socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Socio deleted successfully"
                    },
                    "404": {
                        "description": "Socio not found"
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Register a new user",
                "description": "Register a new user and return an access token",
                "operationId": "b5e5cdd89c27497d47288c39cf2dd8c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "401": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "User login",
                "description": "Login with email and password and return an access token",
                "operationId": "f854efad96f05200354af27e2598a134",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "accessToken"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Name or password incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "User logout",
                "description": "Revoke the user's access token to logout",
                "operationId": "f5ec7320c91d7e1b2b0d16d4e59967a5",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "0": []
        }
    }
}